requirements.txt
django==4.1.7
djangorestframework==3.14.0
python-dotenv==0.21.1
Pillow==9.4.0
django-filter==22.1
djangorestframework_simplejwt
django-redis==5.2.0
gunicorn==20.1.0
django-import-export==3.1.0
django-cors-headers==3.14.0
requests